requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
requests==2.31.0
Pillow==10.1.0
stripe==7.10.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
bcrypt==4.1.2
email-validator==2.1.0.post1
flask-admin==1.6.1
